pyproject.toml
[project]
name = "invite-review"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.50.0",
    "flask>=3.1.0",
    "python-dotenv>=1.1.0",
]
